+++ src/Ghostscript/GhostscriptServiceProvider.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of Ghostscript-PHP.
+ *
+ * (c) Alchemy <info@alchemy.fr>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Ghostscript;
 
 use Silex\ServiceProviderInterface;
@@ -9,23 +18,19 @@ class GhostscriptServiceProvider implements ServiceProviderInterface
 {
     public function register(Application $app)
     {
-        $app['ghostscript.binary'] = null;
+        $app['ghostscript.default.configuration'] = array(
+            'gs.binaries' => array('gs'),
+            'timeout'     => 60,
+        );
+        $app['ghostscript.configuration'] = array();
         $app['ghostscript.logger'] = null;
-        $app['ghostscript.timeout'] = 0;
 
         $app['ghostscript.transcoder'] = $app->share(function(Application $app) {
+            $app['ghostscript.configuration'] = array_replace(
+                $app['ghostscript.default.configuration'], $app['ghostscript.configuration']
+            );
 
-            if (null !== $app['ghostscript.logger']) {
-                $logger = $app['ghostscript.logger'];
-            } else {
-                $logger = null;
-            }
-
-            if (null === $app['ghostscript.binary']) {
-                return Transcoder::create($logger, array('timeout' => $app['ghostscript.timeout']));
-            } else {
-                return Transcoder::load($app['ghostscript.binary'], $logger, array('timeout' => $app['ghostscript.timeout']));
-            }
+            return Transcoder::create($app['ghostscript.configuration'], $app['ghostscript.logger']);
         });
     }